Separate and mark chipper tests
